fefactor: change structure database
